inline.h: Macroize DFA for isFOO_UTF8_CHAR()
There are currently three functions for variants of finding if the next
few bytes of a string form a proper UTF-8 encoded character of some ilk.

The main code for each is identical to the others, except for the table
that drives it.

This commit makes that code a macro that takes arguments to customize
its behavior sufficiently for current and forseeable needs.

This makes it easier to keep the varieties in sync with each other with
future changes.

The macro has three exit points: 1) successful parsing
                                 2) unsuccessful parsing
                                 3) succesful parsing as far as it went,
                                    but the input was exhausted before
                                    reaching a full character.

What to do for each of these eventualities is passed to the macro.  This
is a change in behavior in which 2) and 3) were not distinguished from
each other.  This actually leads to fewer tests in some situations, and
future commits using this DFA for other purposes will take advantage of
it.