handy.h: Fix internal macro
I found this reading code.  The macro is supposed to check for something
not being in the ASCII range, but instead checked that the input is
invariant under UTF-8.  These concepts evaluate to the same thing on
ASCII platforms, but differently on EBCDIC ones.  The calls to this
macro are such that there isn't a bug that surfaces here, but the code
generated is slightly different, and it should be fixed to prevent any
future issues.