utf8.h: Add new #define for extended length UTF-8
The previous commit added a convenient place to create a symbol to
indicate that the UTF-8 on this platform includes Perl's nearly-double
length extension.  The platforms this isn't needed on are 32-bit ASCII
ones.  This symbol allows removing one place where EBCDIC need
be considered, and future commits will use it as well.