MACH-O can't support C11 thread local in shared objects, so disable it.
Apple's clang happily passes the probe and compiles _Thread_local no problem,
but building extensions fails with
    ld: illegal thread local variable reference to regular symbol _PL_current_context for architecture arm64

The Internet suggests that the MACH-O format fundamentally can't support
what is needed to implement C11 thread local storage in shared objects.
It's frustrating that the error message is "Less Than Awesome" at explaining
that this is the real problem here.

Hence disable the use of C11 thread local storage in the hints file, and
hence keep using pthreads.