fix new libperl.t 'no symbols' warning on darwin
A change in October 2019 added handling for the case that 'nm' might
emit the error message 'no symbols' (apparently erroneously. Apple seems
to have included a fix to upstream LLVM to change the message to be a
file reference followed by ': no symbols'. For libperl.t, the message is
typically '../libperl.a:perlapi.o: no symbols'. This change handles both
the old and current behavior.