Add helper function for longest UTF8 sequence
This specialized functionality is used to check the validity of Perl's
extended-length UTF-8, which has some ideosyncratic characteristics from
the shorter sequences.  This means this function doesn't have to
consider those differences.  It will be used in the next commit to avoid
some work, and to eventually enable is_utf8_char_helper() to be
simplified.