Three small changes to perlintro.pod
Simplify the text about declaring variables with my. The previous text said
both that "you don't have to pre-declare" and "you have to declare them",
where it wasn't clear that the former was about types, and the latter about
variables. The new phrasing tries to be clearer.

Note that special variables can also be "all capitals", not just
"punctuation". Remove the in joke about "line noise" because it adds little,
but many people use it as a negative trope about Perl.

Remove mention of given/when. We should not be promoting something we
intend to change. We should have done this years ago when we made it
experimental.