regcharclass.h: #defines for non-chars by UTF8 length
This creates macros for the non-character code points so that, given the
length of the UTF-8 sequence, only those ones that have that length
match.  This makes for more efficient processing, to be used in a future
commit.

The place where the length changes depends on the platform type, and
these macros will keep the code from having to worry about that.