Remove some cruft from XS::APItest.
We no longer need the 5.9.x era work around for UNITCHECK.

The use 5.011001 can go too.  As XS::APItest is only in the core tree and
never installed, the only perl it will ever run with is the version that it
shipped with.

This line was added in Sep 2010 by commit 83f8bb564961ee8f:

    merge XS-APItest-KeywordRPN into XS-APItest

    XS-APItest-KeywordRPN has turned out to be less useful as an independent
    module than expected, and less strictly about RPN than it originally was.

It was previously in the source for XS::APItest::KeywordRPN, and it's unclear
why it was needed there.