Move all the code that deletes the hash entry into one place
Move the code that removed the entry from the hash ahead of all the code
that post-processes that entry to update method caches.

The entry is now completely deleted from the hash before any needed MRO
related logic is triggered, and the code that is specific to the chosen hash
table implementation is now untangled from the MRO code.