use CLANG_DIAG_IGNORE_STMT instead of GCC_DIAG_IGNORE_STMT
It turns out gcc wasn't warning on this code, but older gcc (as
included in debian buster) *does* warn on the switch which it
doesn't recognise.

Newer gcc does recognise the -Wstring-compare switch, but it controls
warning on a different construct, so there's no reason to present it
to gcc.