add MakeMaker test file to  'TEST -deparse' skips
On non-VMS platforms, the cpan/ExtUtils-MakeMaker/t/MM_VMS.t test script
does a plan('skip_all') in a BEGIN block, which stops the rest of the
test file from being compiled. It doesn't matter so much that the rest
of the test file doesn't appear appear in the Deparse output (since we
don't want to execute it anyway); but recently, @Methods was changed to
be a lexical variable, and its 'my' declaration isn't being output.