Test the interaction of of tie/untie and hash iterators
This is not intended as a test of *correctness*. This behaviour is
observable by code on CPAN, so potentially some of it will inadvertently be
relying on it (and likely not in any regression test). Hence this "test"
here is intended as a way to alert us if any core code change has the side
effect of alerting this observable behaviour.