Two minor Makefile.SH cleanups
Use "$osname" instead of (bare) $osname in all case statements.
This was the only one that was different.

Delete PATH_SEP, which was never used.
This was added in Nov 1995 as part of commit 4633a7c4bad06b47:
    5.002 beta 1

with various other macros, and the comment "These variables will be used in
a future version to make the make file more portable to non-unix systems."

The other macros have been used, but not PATH_SEP.

Curiously it also gets a different value if one (re)generates Makefile by
running ./Makefile.SH, compared with a Makefile generated by Configure,
because the variable that it is set from - $p_ - is not in config.sh